code deduplication - actioned too fast
