+++ module/gaminimas_amatu_potion/gaminimas_amatu_potion.go
@@ -2,7 +2,6 @@ package gaminimas_amatu_potion
 
 import (
 	"errors"
-	"log"
 	"net/url"
 	"strings"
 	"tobot/module"
@@ -109,12 +108,9 @@ func (obj *GaminimasAmatuPotion) Perform(p *player.Player, settings map[string]s
 		return &module.Result{CanRepeat: true, Error: nil}
 	}
 
-	// If actioned too fast
-	if doc.Find("div:contains('Jūs pavargęs, bandykite vėl po keleto sekundžių..')").Length() > 0 {
-		log.Println("actioned too fast, retrying...")
+	if module.IsActionTooFast(doc) {
 		return obj.Perform(p, settings)
 	}
-
 	module.DumpHTML(doc)
 	return &module.Result{CanRepeat: false, Error: errors.New("unknown error occurred")}
 }
